requirements.txt
future~=0.18.3
pypdf2~=2.12.1
fitz~=0.0.1.dev2
pytesseract~=0.3.10
Pillow~=9.4.0